Forwarding system complete
+++Added a forwarding flag to the sendData function
===Got forwarding working for both the sendRequest and sendData functions
>>>Tested and the system forwarded to the next avalible available hub as designed
>>>There where hickups in which a hub would freeze up durring forwarding, but they were random. Wont be able to debug until I am able to see the logs during operation
===Work will now be put towards the home sever as I feel I reached 75% completion with the sensor hub

+++ SensorHub/hub.cpp
@@ -218,7 +218,7 @@ void Hub::debugPrint(String message)
 	}
 }
 
-void Hub::debugPrint(float data)
+void Hub::debugPrint(int data)
 {
 	//If debugging has been enabled
 	if (debugMode == true)
